Rename llm-workflows to workflows-llm and drop eslintrc exclusion
Renaming src/llm-workflows.ts to src/workflows-llm.ts lets it match the
shared workflows-*.ts eslint glob, so workflow-streams no longer needs a
custom .eslintrc.js or an entry in ESLINTRC_EXCLUDE.
